// lib/protobuf/PushDataV3ApiWrapper.ts
export interface PushDataV3ApiWrapper {
  /** 频道 */
  channel: string;
  publicDeals?: PublicDealsV3Api | undefined;
  publicIncreaseDepths?: PublicIncreaseDepthsV3Api | undefined;
  publicLimitDepths?: PublicLimitDepthsV3Api | undefined;
  privateOrders?: PrivateOrdersV3Api | undefined;
  publicBookTicker?: PublicBookTickerV3Api | undefined;
  privateDeals?: PrivateDealsV3Api | undefined;
  privateAccount?: PrivateAccountV3Api | undefined;
  publicSpotKline?: PublicSpotKlineV3Api | undefined;
  publicMiniTicker?: PublicMiniTickerV3Api | undefined;
  publicMiniTickers?: PublicMiniTickersV3Api | undefined;
  publicBookTickerBatch?: PublicBookTickerBatchV3Api | undefined;
  publicIncreaseDepthsBatch?: PublicIncreaseDepthsBatchV3Api | undefined;
  publicAggreDepths?: PublicAggreDepthsV3Api | undefined;
  publicAggreDeals?: PublicAggreDealsV3Api | undefined;
  publicAggreBookTicker?:
    | PublicAggreBookTickerV3Api
    | undefined;
  /** 交易对 */
  symbol?:
    | string
    | undefined;
  /** 交易对ID */
  symbolId?:
    | string
    | undefined;
  /** 消息生成时间 */
  createTime?:
    | number
    | undefined;
  /** 消息推送时间 */
  sendTime?: number | undefined;
}